Seperated server code Added folders and test files for each new block Added new redis unit tests
