--- server/src/routes.js
@@ -1,1 +1,0 @@
-// TODO, un dossier route !!
